// src/app/models/cars.ts
export interface Cars {
    ativo?: number;
    nome?: string;
    anho_fabricacao?: string;
    modelo?: string;
    cambio?: number;
    km?: number;
    placa?: string;
    cor?: number;
    carroceria?: string;
    portas?: string;
    motorizacao?: string;
    combustivel?: string;
    chassi?: string;
    valor?: string;
}